requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
Django==5.0.3
gcloud==0.18.3
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.4
oauth2client==4.1.3
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.1.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
